+++ flickit-assessment-core/src/test/resources/postman/flickit-assessment-core.postman_collection.json
@@ -2190,7 +2190,7 @@
 							"response": []
 						},
 						{
-							"name": "Get Assessment Insight",
+							"name": "Get Assessment Overall Insight",
 							"protocolProfileBehavior": {
 								"disableBodyPruning": true
 							},
